+++ packages/filepicker/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.23.12](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.23.11...@standardnotes/filepicker@1.23.12) (2022-10-06)
+
+**Note:** Version bump only for package @standardnotes/filepicker
+
 ## [1.23.11](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.23.10...@standardnotes/filepicker@1.23.11) (2022-10-05)
 
 **Note:** Version bump only for package @standardnotes/filepicker