+++ packages/files/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.12](https://github.com/standardnotes/app/compare/@standardnotes/files@1.10.11...@standardnotes/files@1.10.12) (2022-10-06)
+
+**Note:** Version bump only for package @standardnotes/files
+
 ## [1.10.11](https://github.com/standardnotes/app/compare/@standardnotes/files@1.10.10...@standardnotes/files@1.10.11) (2022-10-05)
 
 **Note:** Version bump only for package @standardnotes/files