+++ packages/services/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.26.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.25.0...@standardnotes/services@1.26.0) (2022-10-07)
+
+### Features
+
+* **api:** add workspaces api ([#1765](https://github.com/standardnotes/app/issues/1765)) ([01ba715](https://github.com/standardnotes/app/commit/01ba715eba987a7da1ee062fec0b3593a7a453ed))
+
+# [1.25.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.24.3...@standardnotes/services@1.25.0) (2022-10-06)
+
+### Features
+
+* experimental 005 operator ([#1753](https://github.com/standardnotes/app/issues/1753)) ([cbbe913](https://github.com/standardnotes/app/commit/cbbe913cd6eb694dd27997927bd5c45e8a64cc09))
+
 ## [1.24.3](https://github.com/standardnotes/app/compare/@standardnotes/services@1.24.2...@standardnotes/services@1.24.3) (2022-10-05)
 
 **Note:** Version bump only for package @standardnotes/services